#000-Update invoice management related features
Invoices now use a Guid instead of an integer for their ID, to enhance uniqueness and avoid potential merging conflicts. Added price and description properties to InvoiceLines, and removed total as it is now computed from price and quantity. It's identified a need for TryCatch functionality in the Result class to handle potential exceptions in a streamlined manner. Restructured several models for cleaner code formatting. Specifically, additional IDatabaseSeeder interface method removed as unnecessary. A new CreateInvoiceCommand was introduced to support the creation of invoices within the application. The Invoice.Create and InvoiceLine.Create are modified to add validation and return possible validation errors."