Add Invoicing functionalities and improve code organization
Added several new functionalities related to creating an invoice. This includes the ability to search for services and routes, add them, and view them in a list. Also added necessary response models related to the invoice. Improved the readability of the code by breaking down long lines of code and creating dedicated components for various functionalities. Also, implemented a try-catch block in the Result.cs to handle exceptions during any request executions. Finally, other minor fixes are also included, like the removal of an unused variable from the Customer.cs and