Update route attributes and add migrations folder in project
This commit removes the `Distance` and `EstimatedTime` attributes from Route objects in the application, as they are no longer considered necessary for the application's functionality. It also changes the migrations directory in the `README.md` file to `Persistences/Migrations` for better organization and understanding of the project structure. Similarly, an 'output-dir' attribute has been added for command 'dotnet ef', specifying the new location. All impacted components and endpoints have been adjusted according to these changes. Additionally, the commit includes the addition of statuses